add crud para usuario, resposta e curso

+++ src/main/java/br/com/forumhub/demo/repository/RespostaRepository.java
@@ -4,6 +4,8 @@ import br.com.forumhub.demo.model.Resposta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface RespostaRepository extends JpaRepository<Resposta, Integer> {
+public interface RespostaRepository extends JpaRepository<Resposta, Long> {
 }
